Moved some components out of internal
This commit moves two processors and an error helper outside the
internal package so they can be imported and used by OT service
derivatives.
