+++ processor/nodebatcher/factory.go
